requirements.txt
fastapi
uvicorn
jinja2
docker
PyYAML
python-multipart
uvicorn[standard]
python-dotenv
httpx
